move _calculate_job_distribution to utils.
